Move model alias definition into alias package
