+++ binary-array-ld-lib/src/main/kotlin/net/bald/alias/AliasDefinition.kt
@@ -29,6 +29,9 @@ interface AliasDefinition {
      */
     fun isReferenceProperty(prop: Property): Boolean
 
+    /**
+     * A substitute [AliasDefinition] implementation that represents a null or empty definition.
+     */
     object Empty: AliasDefinition {
         override fun property(identifier: String): Property? {
             return null
