+++ binary-array-ld-lib/src/main/kotlin/net/bald/alias/ModelAliasDefinition.kt
@@ -1,6 +1,5 @@
-package net.bald.model
+package net.bald.alias
 
-import net.bald.alias.AliasDefinition
 import net.bald.vocab.BALD
 import org.apache.jena.rdf.model.*
 import org.apache.jena.vocabulary.DCTerms
