+++ binary-array-ld-lib/src/test/kotlin/net/bald/model/ModelAliasDefinitionTest.kt
@@ -1,5 +1,6 @@
 package net.bald.model
 
+import net.bald.alias.ModelAliasDefinition
 import net.bald.vocab.BALD
 import org.apache.jena.rdf.model.ModelFactory
 import org.apache.jena.rdf.model.ResourceFactory.createProperty
