+++ binary-array-ld-netcdf/src/main/kotlin/net/bald/netcdf/NetCdfAttributeSource.kt
@@ -1,9 +1,6 @@
 package net.bald.netcdf
 
 import net.bald.AttributeSource
-import net.bald.alias.AliasDefinition
-import org.apache.jena.rdf.model.Property
-import org.apache.jena.shared.PrefixMapping
 import ucar.nc2.Attribute
 import ucar.nc2.AttributeContainer
