+++ binary-array-ld-netcdf/src/main/kotlin/net/bald/netcdf/NetCdfSubContainer.kt
@@ -2,7 +2,6 @@ package net.bald.netcdf
 
 import net.bald.Container
 import net.bald.alias.AliasDefinition
-import net.bald.context.ModelContext
 import ucar.nc2.Group
 
 /**
